create DateFnsLocalizedComponent for shared class functionality
